Split app dataloaders
