+++ saleor/graphql/app/dataloaders/app_extension.py
@@ -1,0 +1,29 @@
+from collections import defaultdict
+
+from ....app.models import AppExtension
+from ...core.dataloaders import DataLoader
+
+
+class AppExtensionByIdLoader(DataLoader):
+    context_key = "app_extension_by_id"
+
+    def batch_load(self, keys):
+        extensions = AppExtension.objects.using(self.database_connection_name).in_bulk(
+            keys
+        )
+        return [extensions.get(key) for key in keys]
+
+
+class AppExtensionByAppIdLoader(DataLoader):
+    context_key = "app_extension_by_app_id"
+
+    def batch_load(self, keys):
+        extensions = AppExtension.objects.using(self.database_connection_name).filter(
+            app_id__in=keys
+        )
+        extensions_map = defaultdict(list)
+        app_extension_loader = AppExtensionByIdLoader(self.context)
+        for extension in extensions.iterator():
+            extensions_map[extension.app_id].append(extension)
+            app_extension_loader.prime(extension.id, extension)
+        return [extensions_map.get(app_id, []) for app_id in keys]
