+++ saleor/graphql/app/types.py
@@ -23,6 +23,7 @@ from ..account.utils import is_owner_or_has_one_of_perms
 from ..core import ResolveInfo, SaleorContext
 from ..core.connection import CountableConnection
 from ..core.context import get_database_connection_name
+from ..core.dataloaders import DataLoader
 from ..core.descriptions import (
     ADDED_IN_31,
     ADDED_IN_35,
@@ -54,7 +55,6 @@ from .dataloaders import (
     AppByIdLoader,
     AppExtensionByAppIdLoader,
     AppTokensByAppIdLoader,
-    DataLoader,
     ThumbnailByAppIdSizeAndFormatLoader,
     ThumbnailByAppInstallationIdSizeAndFormatLoader,
     app_promise_callback,
